invocation and resource usage connected
