+++ backend/internal/adapters/primary/web/server.go
@@ -27,8 +27,7 @@ type Server struct {
 
 func StartServer(db *postgres.Database, mq ports.IMessageQueue, containerMan ports.IContainerManager, fileRepo ports.IFileRepository) {
 	logService := log.NewLogService(mq, db)
-	resourceService := analytics.NewResourceService(db)
-	containerService := container.NewContainerService(containerMan, logService, fileRepo, db, resourceService)
+	containerService := container.NewContainerService(containerMan, logService, fileRepo, db)
 	projectService := project.NewProjectService(db, containerService, fileRepo)
 	envService := envs.NewEnvService(db, *projectService)
 	analyticsService := analytics.NewAnalyticsService(db, mq)
