+++ backend/internal/core/services/container/access_resource.go
@@ -1,0 +1,7 @@
+package container
+
+import "pulsar/internal/core/domain/analytics"
+
+func (cs *containerService) AccessResource() *analytics.RuntimeResourceObj {
+	return cs.resource
+}
