+++ backend/internal/core/services/container/start_app.go
@@ -3,6 +3,7 @@ package container
 import (
 	"context"
 	"fmt"
+	"pulsar/internal/core/domain/analytics"
 	domain "pulsar/internal/core/domain/log"
 	"pulsar/internal/core/domain/project"
 	"pulsar/internal/core/services"
@@ -30,6 +31,9 @@ func (cs *containerService) startServerlessApp(ctx context.Context, project *pro
 			))
 			errorChan <- services.NewAppError(services.ErrInternalServer, err)
 		}
+
+		cs.resource = analytics.NewRuntimeResObj()
+		cs.monitor = analytics.NewRuntimeResMonitor()
 
 		go cs.containerMan.GetContainerStats(ctx, project.ContainerId, cs.resource, cs.monitor)
