+++ backend/internal/ports/invocation_repository.go
@@ -10,4 +10,6 @@ type InvocationRepository interface {
 	GetInvocationsOfLast24Hours(ctx context.Context, projectId string, status analytics.InvocationStatus) ([]*analytics.InvocationCount, error)
 	GetInvocationsOfLast7Days(ctx context.Context, projectId string, status analytics.InvocationStatus) ([]*analytics.InvocationCount, error)
 	GetInvocationsOfLast30Days(ctx context.Context, projectId string, status analytics.InvocationStatus) ([]*analytics.InvocationCount, error)
+	CreateResourceUtil(ctx context.Context, resource *analytics.RuntimeResource) error
+	GetInvocationResourceUtil(ctx context.Context, utilizationId string) (*analytics.RuntimeResource, error)
 }
